- Timer/Counter parameter type 변경: uint16 -> uint32

+++ DsDotNet/src/Engine/Engine.CodeGenCPU/Core.Extensions/DsSystem.fs
@@ -62,7 +62,7 @@ module ConvertCpuDsSystem =
         member s._homingMonitor = s.GetPv<bool>(SystemTag.homingMonitor  )
         member s._goingMonitor  = s.GetPv<bool>(SystemTag.goingMonitor  )
 
-        member s._tout        = s.GetPv<uint16>(SystemTag.timeout)
+        member s._tout        = s.GetPv<uint32>(SystemTag.timeout)
         member s._flicker20msec = s.GetPv<bool>(SystemTag.flicker20ms)
         member s._flicker100msec = s.GetPv<bool>(SystemTag.flicker100ms)
         member s._flicker200msec = s.GetPv<bool>(SystemTag.flicker200ms)
